feature: commission_rate_in_basis_points (#9822)

+++ src/flamenco/runtime/fd_runtime_stack.h
@@ -32,7 +32,7 @@ struct fd_vote_rewards {
   fd_pubkey_t pubkey;
   ulong       vote_rewards;
   uint        next;
-  uchar       commission;
+  ushort      commission;
 };
 typedef struct fd_vote_rewards fd_vote_rewards_t;
